requirements.txt
django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.11.0
psycopg2-binary==2.9.3
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2
